support vector operand bcst,{sae},{er} + k-register
git-svn-id: branches/tg74/avx512@39457 -

+++ compiler/aasmtai.pas
@@ -420,7 +420,7 @@ interface
         toper = record
           ot : longint;
         {$ifdef x86}
-          vopext: byte;
+          vopext: smallint;
         {$ENDIF}
           case typ : toptype of
             top_none   : ();
